requirements.txt
click==7.1.2
PyYAML==5.3.1
requests==2.22.0
Pillow==7.0.0
rich==6.1.2
